Fix for #3325. localizableStrings is no longer an array
