+++ lib/devices/ios/ios-controller.js
@@ -256,8 +256,8 @@ iOSController.getLocalizedStringForSelector = function (selector) {
   var newSelector = selector;
 
   var strings = this.localizableStrings;
-  if (strings && strings.length >= 1) {
-    var localizedSelector = strings[0][selector];
+  if (strings) {
+    var localizedSelector = strings[selector];
     if (localizedSelector) {
       newSelector = localizedSelector;
     } else {
